Fix two bugs when moving matrices to the host
There was a bug where the IndexSet of the domain and the range where inverted
in the constructor. The second bug is in the convert_matrix of
dealii::SparseMatrix. When the matrix is rectangular the function diag_element
does not exist and there is no need to reorder the underlying data.